memory/short_term_memory_example/requirements.txt
google-adk
google-genai
